+++ cdl/core/baseproxy.py
@@ -284,6 +284,18 @@ class AbstractCDLControl(abc.ABC):
         """
 
     @abc.abstractmethod
+    def add_group(
+        self, title: str, panel: str | None = None, select: bool = False
+    ) -> None:
+        """Add group to DataLab.
+
+        Args:
+            title: Group title
+            panel: Panel name (valid values: "signal", "image"). Defaults to None.
+            select: Select the group after creation. Defaults to False.
+        """
+
+    @abc.abstractmethod
     def get_sel_object_uuids(self, include_groups: bool = False) -> list[str]:
         """Return selected objects uuids.
 
@@ -621,6 +633,18 @@ class BaseProxy(AbstractCDLControl, metaclass=abc.ABCMeta):
             List of selected objects uuids.
         """
         return self._cdl.get_sel_object_uuids(include_groups)
+
+    def add_group(
+        self, title: str, panel: str | None = None, select: bool = False
+    ) -> None:
+        """Add group to DataLab.
+
+        Args:
+            title: Group title
+            panel: Panel name (valid values: "signal", "image"). Defaults to None.
+            select: Select the group after creation. Defaults to False.
+        """
+        self._cdl.add_group(title, panel, select)
 
     def select_objects(
         self,